move flow folder, fix tool warning and fix progress bar
