+++ examples/gen_test_data/generate_test_data_flow/generate_debug_info.py
@@ -1,6 +1,6 @@
 from utils import ValidateObj, ValidationResult
 
-from promptflow import tool
+from promptflow.core import tool
 
 
 # The inputs section will change based on the arguments of the tool function, after you save the code
