+++ examples/gen_test_data/generate_test_data_flow/validate_question.py
@@ -2,9 +2,9 @@ from typing import Union
 
 from utils import ErrorMsg, QuestionType, ResponseFormat, get_question_validation_res
 
-from promptflow import tool
 from promptflow._core.tool import InputSetting
 from promptflow.connections import AzureOpenAIConnection, OpenAIConnection
+from promptflow.core import tool
 
 
 @tool(
