+++ examples/gen_test_data/generate_test_data_flow/validate_question_prompt.jinja2
@@ -3,7 +3,7 @@
 Verdict a question based on following rules:
 
 1. If there are acronyms or terms in the question, then please check if they exist in the given context. If no, verdict no. If yes, check if other rules are satisfied.
-2. Determine if the given question can be clearly understood and give the reason. 
+2. Determine if the given question can be clearly understood and give the reason.
 
 Output a valid json with reason and verdict.
