+++ examples/gen_test_data/requirements_cloud.txt
@@ -1,4 +1,4 @@
 promptflow>=1.7.0
 promptflow-tools
-azure-ai-ml
-mldesigner
+azure-ai-ml==1.15.0
+mldesigner==0.1.0b18
